Use unsigned chars, probably a good idea.
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/branches/pocketsphinx-0.5@8548 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ pocketsphinx/src/libpocketsphinx/bin_mdef.c
@@ -123,9 +123,9 @@ bin_mdef_read_text(cmd_ln_t *config, const char *filename)
         }
         else {
             bmdef->phone[i].info.cd.wpos = mdef->phone[i].wpos;
-            bmdef->phone[i].info.cd.ctx[0] = mdef->phone[i].ci;
-            bmdef->phone[i].info.cd.ctx[1] = mdef->phone[i].lc;
-            bmdef->phone[i].info.cd.ctx[2] = mdef->phone[i].rc;
+            bmdef->phone[i].info.cd.ctx[0] = (uint8)mdef->phone[i].ci;
+            bmdef->phone[i].info.cd.ctx[1] = (uint8)mdef->phone[i].lc;
+            bmdef->phone[i].info.cd.ctx[2] = (uint8)mdef->phone[i].rc;
         }
     }
